Module system now checks for unused dependencies.
Note: Nodes, IOInterfaces, and DataTypes are still being directly
assigned to properties of the Synesthesia object which causes the
module checker to show them as unused dependencies. These modules
need to be made side-effect free and requested using the
module.require method instead of just asking for them in a module
declaration and taking their results from the places they are
assigned to on other objects.