Removed code for getting MIDIAccess from main context.
Should MIDI really be handled separately by each MIDISource node
or should the main context manage a single MIDIAccess object and
MIDISource objects request inputs and outputs from the main context?
This commit leans towards the former, but as a whole I'm still not
entirely sure which is a better design. Having a single MIDIAccess
object might make it simpler to manage the devices in one place and
would parallel the need to have a single audio context.